requirements.txt
discord-py-paginators
discord.py
aiohttp
humanize